Add pages for signin and signup instead of modals
