+++ app/components/application/index.jsx
@@ -1,5 +1,4 @@
 import React, { Component } from 'react';
-import Modals from 'containers/modals';
 import Header from 'components/header';
 import Footer from 'components/footer';
 import styles from './styles';
@@ -13,7 +12,6 @@ export default class Application extends Component {
           { this.props.children }
         </main>
         <Footer/>
-        <Modals/>
       </div>
     );
   }
