+++ app/components/header/index.jsx
@@ -1,16 +1,11 @@
 import React, { Component } from 'react';
 import { Navbar } from 'react-bootstrap';
-import NavigationLeft from 'components/navigation/left';
-import NavigationRight from 'components/navigation/right';
 import { paths } from 'helpers/routes';
+import Navigation from 'components/navigation';
+import NavigationRight from 'components/navigation/right';
 
 export default class Header extends Component {
   render() {
-    const links = [
-      { title: 'Home', route: paths.home() },
-      { title: 'About', route: paths.about() }
-    ];
-
     return (
       <Navbar>
         <Navbar.Header>
@@ -18,7 +13,7 @@ export default class Header extends Component {
             React-base
           </Navbar.Brand>
         </Navbar.Header>
-        <NavigationLeft items={ links }/>
+        <Navigation/>
         <NavigationRight/>
       </Navbar>
     );
