+++ app/helpers/routes.js
@@ -3,6 +3,8 @@ import session from 'services/session';
 export const paths = {
   home() { return '/'; },
   about() { return '/about'; },
+  signin() { return '/signin'; },
+  signup() { return '/signup'; },
   aboutExtended(id) { return `/about/extended/${id}`; }
 };
