+++ app/stores/session.js
@@ -1,7 +1,6 @@
 import Alt from 'altFlux';
 import { createStore } from 'alt-utils/lib/decorators';
 import Storage from 'lib/storage';
-import SessionActions from 'actions/session';
 import config from 'config';
 
 const STORAGE_KEY = config.storageKey;
@@ -12,11 +11,6 @@ export default class SessionStore {
 
   constructor() {
     this.currentUser = Storage.get(STORAGE_KEY) || {};
-
-    this.bindListeners({
-      create: SessionActions.CREATE,
-      delete: SessionActions.DELETE
-    });
   }
 
   create(user) {
